feat - Dialogos de confirmacion y validacion de producto (#281)
* feat - Dialogos de confirmacion y validacion de producto

Closes #175

* feat - Dialogos responsivos

- Se implementa que los dialogos se posicionen en la parte inferior en mobile y central en desktop
- Se refactoriza codigo para hacerlo mas limpio.

+++ lib/modulos/productos/ui/vista_modificar_producto.dart
@@ -1,4 +1,5 @@
 import 'package:eleventa/modulos/common/ui/ex_icons.dart';
+import 'package:eleventa/modulos/common/ui/widgets/ex_dialogos.dart';
 import 'package:eleventa/modulos/common/ui/widgets/ex_vista_principal_scaffold.dart';
 import 'package:eleventa/modulos/common/utils/uid.dart';
 import 'package:eleventa/modulos/productos/modulo_productos.dart';
@@ -9,11 +10,11 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:eleventa/modulos/productos/ui/listado_productos_provider.dart';
 import 'package:eleventa/modulos/productos/ui/forma_producto.dart';
 
-class ModificarProducto extends StatelessWidget {
+class VistaModificarProducto extends StatelessWidget {
   static const titulo = 'Modificar Producto';
   final String productoId;
 
-  const ModificarProducto({
+  const VistaModificarProducto({
     Key? key,
     required this.productoId,
   }) : super(key: key);
@@ -32,32 +33,39 @@ class ModificarProducto extends StatelessWidget {
   @override
   Widget build(BuildContext context, [bool mounted = true]) {
     return VistaPrincipalScaffold(
-        titulo: ModificarProducto.titulo,
+        titulo: VistaModificarProducto.titulo,
         actions: [
           Consumer(
             builder: (context, ref, child) => IconButton(
               icon: const Icon(Iconos.trash, color: Colors.white),
               tooltip: 'Eliminar producto',
               onPressed: () async {
-                await _eliminarProducto();
+                var eliminarProducto =
+                    await ExDialogos.mostrarConfirmacionEliminar(context,
+                        titulo: '¿Eliminar producto?');
 
-                if (!mounted) return;
+                if ((eliminarProducto != null) && (eliminarProducto == true)) {
+                  await _eliminarProducto();
+                  if (!mounted) return;
 
-                // Solo aplicar en mobile
-                if (context.breakpoint <= LayoutBreakpoint.sm) {
-                  context.pop();
+                  // Solo aplicar en mobile
+                  if (context.breakpoint <= LayoutBreakpoint.sm) {
+                    context.pop();
+                  }
+
+                  ref
+                      .read(providerListadoProductos.notifier)
+                      .obtenerProductos();
+
+                  ScaffoldMessenger.of(context).showSnackBar(
+                      const SnackBar(content: Text('Producto eliminado')));
                 }
-
-                ref.read(providerListadoProductos.notifier).obtenerProductos();
-
-                ScaffoldMessenger.of(context).showSnackBar(
-                    const SnackBar(content: Text('Producto eliminado')));
               },
             ),
           )
         ],
         child: FormaProducto(context,
             productoEnModificacionId: productoId,
-            titulo: ModificarProducto.titulo));
+            titulo: VistaModificarProducto.titulo));
   }
 }
